+++ CHANGELOG.md
@@ -6,6 +6,7 @@ ChangeLog
 - fix(js): fix #2439, `filterBy` cannot filter array keys
 - fix(js): fix #2424, from html with checkbox bug
 - feat(js): fix #2414, add `paginationLoop` option
+- feat(cookie extension): fix #2386, add `getCookies` method
 
 ### 1.11.0